// src/interfaces/interfaces.ts
export interface CharactersResponse {
  fetchCharacters: FetchCharacters
}

export interface FetchCharacters {
  info: Info
  results: Result[]
}

export interface Info {
  count: number
  pages: number
  next: string
  prev: string
}

export interface Result {
  id: number
  name: string
  status: string
  species: string
  type: string
  gender: string
  origin: Location
  location: Location
  image: string
  episode: string[]
  url: string
  created: string
}

export interface Location {
  name: string
  url: string
}